Extend the ListView selection behind the checkbox.
Extend the selection behind the checkbox and change the behavior of selecting an
alarm item. Long pressing the item now brings up a context menu that can enable
or disable the alarm in addition to deleting the alarm.

The context menu header now contains the label of the alarm if it exists and
it mimics the standard dialog header view.

Updated some of the layout code to ensure that the default three Alarms do not
show a scrollbar.

BUG=1438269