Allow the alert to be silent.
Turn on the silent item in the ringtone picker. Detect the null alert and use a
special "silent" string as the alert to detect a silent alert. Update the
summary of the alert to show Silent as the ringtone.

Change the SetAlarm manifest to get config changes for orientation and keyboard.
This prevents the settings from resetting when changing the orientation.