Huge rewrite of the AlarmClock to play the Alarm in a service.
The AlarmKlaxon has been converted to a service that plays the alarm and
vibrates the device. The AlarmAlert now just shows the UI for the alarm and
allows the user to snooze or dismiss the alarm. The snooze button must be
pressed in order to snooze the alarm. Volume and Camera buttons dismiss the
alarm while other buttons have their original behavior.

Each alarm triggers a notification that the alarm has fired. This allows another
activity (say, the Calendar alert #1908616) to play on top of the AlarmAlert.
The AlarmKlaxon service will continue to play even though the alert has been
dismissed. The user can get back to the UI through the notification.

If the user snoozes the alarm, the notification reflects that choice (#1691034)
and allows the user to cancel the snooze by clicking the notification. The
snoozed alarm takes priority over any other alarm (#1693155) so that it will
play unless the notification is clicked.

The database interaction has also been rewritten to use a Parcelable Alarm class
for sending and receiving the alarm data in a much simpler manner. This allows
for fewer database lookups since each activity no longer has to lookup the alarm
info.

The alarm silenced text has been removed from the AlarmAlert UI and moved to the
notification area. When an alarm is killed, the alert is dismissed and the
notification reflects the state. Clicking the notification launches the SetAlarm
activity so the user can see which alarm was killed.