Fix issue #2154688: when alarm rings ,device does not wake up...
...and can't see Dismiss and Snooze button if the device is on dock

Change-Id: Ib76b54f0a65930e142a3da1e1152f70d7f3ad3c7
Explanation: only the fullscreen alarm activity had the flags set to turn on
the screen and dismiss the lock screen.  The app picks whether to show
that one or the non-fullscreen one based on whether device is currently
locked, which is correct, but with the introduction of docks, locked
!= screen off.  In particular, the desk dock keeps the keyguard dismissed
while the device is in it.

The fix is to just set these flags for both alert activities.  That seems
generally right for many reasons.