Switch from dialog to fullscreen when the keyguard is showing.
Make AlarmAlert derive from AlarmAlertFullScreen. This looks like a massive
change but it is mostly a rename. Use a broadcast receiver to listen for the
screen turning off. Attempt to read the keyguard status up to 5 times in order
to launch the fullscreen activity. Add "screen_off" to the intent extras to
indicate that the intent is from the screen_off broadcast and to not wake the
screen.

The overall goal of this is that if the alarm fires while the device is on and
the user turns the screen off and then on, the full screen alert will show so
that the user can dismiss the alert.