Completed weather support; new screen-saving mode.
Now pulls high and low temperatures, as well as an icon (in
the form of a drawable resource), from the Genie widget,
periodically.  Known issues: artwork is mdpi and fuzzy (bug
in Genie).

Preliminary support for a new screen-saving mode (subject to
change) specifically for OLEDs with short B/R subpixel
lifetime.  Text is drawn in green on a black background; the
text also randomly jumps around the screen. Any user input
(touch, trackball/keypress, etc.) will restore the previous
view.

Screen saver mode kicks in after a delay (or, for debugging,
a longpress on "Dim").  Honors the current dimness setting,
so if the clock is currently in nightstand mode, the screen
saver will be very dim; otherwise (viz., desk mode) the
saver will be brighter.