UI tweaks to meet Jeff's spec.
Colors and text sizes are exactly how Jeff specified in the pdf. I tried to
match the size of everything based on looks but I might be a little off in
padding. The shadow effect on the bottom clock has been left out for now since
there are no specifics in the spec for each of the shadow components and I
cannot figure it out based on the drawing.