Allow picking ringtones as alarm alerts.
The original request was for just a few of the ringtones but that really isn't
possible with the current framework. Media files are classified by what folder
they are transferred to during the build. It would be nice to have a way to
specify what media files should be allowed in the list of ringtones but for now,
I think it is ok to allow selecting any of the built-in ringtones.