Dismiss the desk clock if it was launched by docking.
In other words: if the clock is behaving like a dock app
(launched by a dock event), it should finish() when the
device is removed from the dock.  If, on the other hand,
it's behaving like a regular app (launched from the
Launcher), it should ignore an un-dock event.

This change also removes support for entering the desk dock
via a dialer code (an unnecessary feature since the app can
always be invoked from the Launcher).

Fixes http://b/2302215, approved by hiroshi.