Have "Revert" go back to the original state of the alarm.
Keep the original alarm state so that revert will reload the preferences from
the original state. Have all preferences immediately take effect. When adding a
new alarm, add the alarm if preferences change and then delete the alarm when
reverting.

Pop the set alarm toast only when changing the time or the enabled setting and
the result is that the alarm is enabled. Otherwise there can be a lot of toasts
showing while editing an alarm.