Add logs to track missed alarms.
Turn on a few verbose logs to track a couple of cases where an alarm might be
skipped or cancelled.  Add a Log.wtf when an alert cannot be parsed as this is
never supposed to happen.

If we are returning from the AlarmReceiver due to some error, make sure to set
the next alert.

Bug: 3004064
Change-Id: I6879317b233ccb0497a5ed7b8259925cd00cf8e8