Fix the design of the alarm popup that wakes the device.
The only way to turn on the screen is to have a fullscreen
theme; doing so disallows us access to the platform styles
for the alert/dialog themes, so this change pulls in a copy
of the correct background bitmap. (We are able, at least,
to use the Holo version of the fullscreen transparent theme,
which causes the buttons to have the correct appearance.)

There's now also a separate layout file for the fullscreen
case so we don't apply this new decoration when a proper
dialog is available to us (read: the screen was already
unlocked).

Finally, tweak AndroidClockTextView so Eclipse doesn't choke
on the layout files.

Bug: 5108144
Change-Id: Ifc153bd9d7a3b9e3d0e18e94ec1b9545976f3759