Fix dropped alarm when snoozing second
When calculating the next alarm event, make sure to take into
account the snoozed previous alarms with the upcoming alarms.

But, now that we allow for another alarm to happen while a previous
one was snoozed, we need to allow for multiple snoozed alarms.

Once we allow for multiple snoozed alarms, we need to be able to allow
for each to be canceled individually.

Bug: 4769415
Change-Id: Icc825a8d892e7b9f53e21499a7af2322b8acbe77