Use dp instead of sp in the alarm alert.
The clock is supposed to be sized to maximally fit the
available space, so there's no point in adjusting for the
user's font size preference.

This change also switches the AM/PM font back to the system
font (and not AndroidClock) to better match the stroke
weight of the numerals, consistent with the DeskClock
display.

Bug: 5446727
Change-Id: Ifb1b65b3041e259e91552b59a61d8219a7b1be94