DeskClock is no longer a "dock home".
It will launch when you dock your ICS device, but it will no
longer take over the home key.

The "dock home" feature still makes sense for, e.g., car
docks where it's appropriate to cripple the device in this
way, but for desktop docks it's just annoying.

In this case it's worse, though, because DeskClock has shed
all of its dock home features (including, notably, a button
to take you to the real Launcher). Without this change, a
docked device is trapped in DeskClock (unless you use
recents or a notification to escape to some other app).

Bug: 5551030
Change-Id: Ib143a33b69b3cb3d092779a8f322e3ed8ae16a84